Improve scaling of MPI parallel sort
The find_splitting_points() routine tries to find the Nth largest
value in a distributed array by bisection, repeatedly estimating
the median over all MPI ranks. When finding multiple splitting
points the median calculation was duplicated on all MPI ranks.

This commit uses MPI_Alltoall to rearrange the data so that each
MPI rank has all of the information to compute the median for one
splitting point. The medians can then be computed in parallel.

This seems to help when we have many MPI ranks sorting a relatively
small array.