Workaround for crash in MultiFile.write()
Creating a dataset with collective_write() and then adding
attributes with h5file[dataset_name].attrs[...] = ...
sometimes results in

'Unable to open object (unable to modify constant message)'

The problem might be that the dataset is implicitly closed
and reopened by indexing the file. If the h5py.Dataset object
is returned from collective_write() and used to set the
attributes then we don't get a crash.