Add read_attributes option to MultiFile.read()
If read_attributes=True then MultiFile.read() will return an
ndarray subclass with an attribute .attrs, which is a dict
containing any HDF5 attributes of the dataset.

It is assumed that in multi file outputs datasets with the same
name in different files have the same attributes (as is the case
with SWIFT units).