Add an extra test for inconsistent dtypes
The MultiFile.read() test for detecting inconsistent dtypes fails
on one rank because one file can't be inconsistent. This makes it
use more files than ranks so we can cause an error on one rank.

I've also added a test that always uses collective mode, which is
skipped if we're using two or fewer ranks.